requirements.txt
asgiref==3.5.2
certifi==2022.12.7
charset-normalizer==2.1.1
cloudinary==1.30.0
coverage==6.5.0
Django==4.1.4
django-admin-adminlte==1.0.1
django-admin-interface==0.24.2
django-cloudinary-storage==0.3.0
django-colorfield==0.8.0
django-filter==22.1
djangorestframework==3.14.0
idna==3.4
Markdown==3.4.1
Pillow==9.3.0
pytz==2022.7
gunicorn==20.1.0
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.13
whitenoise==6.2.0
